+++ src/Journalist.EventStore/Constants.cs
@@ -1,0 +1,10 @@
+namespace Journalist.EventStore
+{
+    internal class Constants
+    {
+        public class Settings
+        {
+            public const int DEFAULT_EVENT_SLICE_SIZE = 100;
+        }
+    }
+}